extracting out a class for steam
